refactor: migrate core utilities and job management to DIRACCommon
- Move JDL, List utilities from DIRAC.Core.Utilities to DIRACCommon
- Migrate JobManifest and JobDBUtils from WorkloadManagementSystem to DIRACCommon
- Add Pydantic-based JobModel with validation for job descriptions
- Update DIRAC modules to import from DIRACCommon maintaining backward compatibility
- Add diraccfg and pydantic dependencies to support enhanced functionality
