+++ dirac-common/tests/WorkloadManagementSystem/Client/JobState/__init__.py
@@ -1,0 +1,1 @@
+# Init file for WorkloadManagementSystem.Client.JobState tests
