+++ dirac-common/tests/WorkloadManagementSystem/Client/__init__.py
@@ -1,0 +1,1 @@
+# Init file for WorkloadManagementSystem.Client tests
